chore: adiciona sharding
