+++ src/index.js
@@ -1,37 +1,23 @@
-const { Client, GatewayIntentBits, Collection, EmbedBuilder } = require("discord.js")
-require("dotenv").config()
-const { loadEvents } = require("./handlers/events.js")
-const { loadCommands } = require("./handlers/commands.js")
+const { ShardingManager } = require("discord.js")
 
-class Falendario {
-	config = require("./config.json")
-	database = require("./handlers/database.js")
+const manager = new ShardingManager("./src/bot.js", {
+	token: process.env.TOKEN,
+	respawn: true,
+	totalShards: 1,
+	shardList: "auto",
+})
 
-	client = new Client({
-		intents: [GatewayIntentBits.Guilds],
+manager.on("shardCreate", (shard) => {
+	shard.on("ready", () => {
+		console.log(`Shard ${shard.id} connected`)
 	})
+})
 
-	constructor() {
-		this.client.on("ready", () => {
-			console.log("Bot online")
-
-			this.client.events = new Collection()
-			this.client.commands = new Collection()
-
-			loadEvents(this, this.client)
-			loadCommands(this, this.client)
-		})
-
-		this.client.login(process.env.TOKEN)
-
-		setInterval(() => {
-			this.client.user.setActivity("Te ajudando a lembrar das coisas importantes", { type: "WATCHING" })
-		}, 1000 * 60 * 1)
-	}
-
-	createEmbed(color = "Random") {
-		return new EmbedBuilder().setColor(color).setFooter({ text: "by Falcão ❤️" })
-	}
-}
-
-Falendario = new Falendario()
+manager
+	.spawn({ amount: manager.totalShards, delay: null, timeout: -1 })
+	.then((shards) => {
+		console.log(`${shards.size} shard(s) spawned.`)
+	})
+	.catch((err) => {
+		console.error("An error has occurred :", err)
+	})
